Modularized code and added airline The pull request contains the following changes:
- Created folders contained the now split up code.

- Created a main.cpp to how global function defintions and main code interface.

- Created the airline class and defined some functions
