+++ README.md
@@ -1,2 +1,4 @@
 # Flight-Management-C-Project
 Real world application of C++ skills
+
+Made by:
